TASK: Remove plow-js usage
The library plow-js has been marked as deprecated and is planned to be removed in Neos 9.0. So we replace the usage of that library in our UI plugins with vanilla JavaScript. In Neos 8.3 plow-js has been deprecated and has been removed in Neos 9.0
So this change removes the dependency and the single usage.

The optional chaining operator was not possible to use here regarding a transpile error, but when the package starts using the esbuild build chain that has been introduced with Neos 8.2 this can be changed.